moved video code from App.js to new VideoDisplay.js component
